Major change to the way shell scripts and patch files are handled.
A new script searches for the correct shell script and patch file for any given install-X or *environment and applies patching if needed. The behavior appears identical to the user, Dockerfiles or deploy scripts can still easily be created by calling "make" or "make deploy" respectively in any distribution directoy. The most noticable difference in the file hierarchy is the lack of symbolic links, they are all removed.